Refactor do diretório da listagem de Chamados de um Equipamento.
